requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-chroma
chromadb
python-dotenv
pypdf
tiktoken
langchain-ollama
